requirements.txt
flask
pyotp
hvac
mysql-connector-python
